update and test rule-based reasoning

+++ ecarf-core/src/main/java/io/ecarf/core/reason/rulebased/owl2rl/rdfs/CaxScoRule.java
@@ -86,7 +86,7 @@ public class CaxScoRule extends GenericRule {
         Object predicate;
 
         if(schemaTriple.isEncoded()) {
-            predicate = SchemaURIType.RDF_TYPE.id;
+            predicate = (long) SchemaURIType.RDF_TYPE.id;
         } else {
             predicate = SchemaURIType.RDF_TYPE.getUri();
         }
